refactor(all): renaming folder, ongoing.
