+++ src/App.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
 import { createGlobalStyle } from 'styled-components';
 import { Reset } from 'styled-reset';
 import Header from './component/Header';
-import { Landing } from './features/Landing/Landing';
+import { Landing } from './features/Xanding/Landing';
 import { Main } from './features/Main/Main';
 import AuthRoute from './component/AuthRoute';
