+++ src/features/Xedger/Amount.tsx
