+++ src/features/Xity/housesSvg/HouseCityHall.tsx
