+++ src/features/Xity/housesSvg/HouseGround.tsx
