+++ src/features/Xity/housesSvg/HouseOfClothes.tsx
