+++ src/features/Xity/housesSvg/HouseOfDrinks.tsx
