+++ src/features/Xity/housesSvg/HouseOfPlants.tsx
