+++ src/features/Xtatistics/GeneralAnalysis.tsx
